Move commands into a commands package
